requirements.txt
albumentations==1.3.0
numpy==1.21.5
pandas==1.4.4
Pillow==9.5.0
torch==2.0.0
torchmetrics==0.11.4
torchvision==0.15.1
tqdm==4.65.0
